move common date/time range manipulation logic to resolutions root
